move visualization into it's own module
